Migrate GroupId to the official
