+++ examples/config_file/pom.xml
@@ -4,7 +4,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.outscale.example</groupId>
+  <groupId>io.github.outscale.example</groupId>
   <artifactId>configfile</artifactId>
   <version>1.0-SNAPSHOT</version>
 
@@ -26,7 +26,7 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>com.outscale</groupId>
+      <groupId>io.github.outscale</groupId>
       <artifactId>osc-sdk-java</artifactId>
       <version>0.1.0</version>
     </dependency>
@@ -57,7 +57,7 @@
           <configuration>
             <archive>
               <manifest>
-                <mainClass>com.outscale.example.App</mainClass>
+                <mainClass>io.github.outscale.example.App</mainClass>
               </manifest>
             </archive>
             <descriptorRefs>
