+++ examples/volumes/src/test/java/io/github/outscale/example/AppTest.java
@@ -1,0 +1,18 @@
+package io.github.outscale.example;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest {
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue() {
+        assertTrue(true);
+    }
+}
