+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.outscale</groupId>
+    <groupId>io.github.outscale</groupId>
     <artifactId>osc-sdk-java</artifactId>
     <packaging>jar</packaging>
     <name>osc-sdk-java</name>
